refactor: extract clearJsPixelData into shared utility
The pattern of clearing JS-side pixel data (delete from pixelDataMap,
sparseMap, add to dirtyIds) was duplicated in 10+ files. Extract into
a single utility function in src/app/store/clear-js-pixel-data.ts and
replace all inline instances.
