+++ src/app/MenuBar/filter-actions.ts
@@ -1,4 +1,5 @@
 import { useEditorStore } from '../editor-store';
+import { clearJsPixelData } from '../store/clear-js-pixel-data';
 import { getEngine } from '../../engine-wasm/engine-state';
 import {
   filterInvert,
@@ -31,16 +32,6 @@ function getActiveLayerId(): string | null {
   return useEditorStore.getState().document.activeLayerId;
 }
 
-function clearJsPixelData(layerId: string): void {
-  const state = useEditorStore.getState();
-  const pixelDataMap = new Map(state.layerPixelData);
-  pixelDataMap.delete(layerId);
-  const sparseMap = new Map(state.sparseLayerData);
-  sparseMap.delete(layerId);
-  const dirtyIds = new Set(state.dirtyLayerIds);
-  dirtyIds.add(layerId);
-  useEditorStore.setState({ layerPixelData: pixelDataMap, sparseLayerData: sparseMap, dirtyLayerIds: dirtyIds });
-}
 
 export function getFilterDialogConfig(id: FilterDialogId): FilterDefinition | null {
   return filterRegistry[id] ?? null;
