+++ src/panels/LayerPanel/layer-selection.ts
@@ -1,5 +1,6 @@
 import { useEditorStore } from '../../app/editor-store';
 import { useUIStore } from '../../app/ui-store';
+import { clearJsPixelData } from '../../app/store/clear-js-pixel-data';
 import { selectionBounds } from '../../selection/selection';
 import { createTransformState } from '../../tools/transform/transform';
 import { getEngine } from '../../engine-wasm/engine-state';
@@ -12,19 +13,7 @@ export function selectLayerAlpha(layerId: string): void {
     dropFloat(engine);
   }
 
-  // Clear stale JS pixel data so resolvePixelData reads from committed GPU texture
-  const preState = useEditorStore.getState();
-  const pixelDataMap = new Map(preState.layerPixelData);
-  pixelDataMap.delete(layerId);
-  const sparseMap = new Map(preState.sparseLayerData);
-  sparseMap.delete(layerId);
-  const dirtyIds = new Set(preState.dirtyLayerIds);
-  dirtyIds.add(layerId);
-  useEditorStore.setState({
-    layerPixelData: pixelDataMap,
-    sparseLayerData: sparseMap,
-    dirtyLayerIds: dirtyIds,
-  });
+  clearJsPixelData(layerId);
 
   const editorState = useEditorStore.getState();
   const layer = editorState.document.layers.find((l) => l.id === layerId);
